created the code structure in front end and deleted the files which was not required created dummy pages for all the links, and added navigations (full functioning needs to be created) counsellor sign up stepper background fix
