+++ my_app/lib/screens/dashboards/userDashboard/chatting_page.dart
@@ -5,7 +5,7 @@ import 'package:flutter/foundation.dart';
 import 'package:flutter/material.dart';
 import 'package:http/http.dart' as http;
 import 'package:firebase_database/firebase_database.dart';
-import 'package:my_app/screens/dashboards/video_player_widget.dart';
+import 'package:my_app/screens/customWidgets/video_player_widget.dart';
 import '../../../services/chat_service.dart';
 import 'package:url_launcher/url_launcher.dart';
 import 'dart:convert'; // For JSON decoding
