--- my_app/lib/screens/oldSignUpScreens/signup_success.dart
@@ -1,79 +1,0 @@
-import 'package:flutter/material.dart';
-import 'package:my_app/screens/oldSignUpScreens/counsellor_signin.dart';
-
-class SignUpSuccessScreen extends StatelessWidget {
-  @override
-  Widget build(BuildContext context) {
-    return Scaffold(
-      backgroundColor: Colors.white,
-      appBar: AppBar(
-        backgroundColor: Color(0xFFFFE4B5),
-        title: Text(
-          "Sign Up Successful",
-          style: TextStyle(color: Colors.white, fontWeight: FontWeight.bold),
-        ),
-        centerTitle: true,
-      ),
-      body: Center(
-        child: Column(
-          mainAxisAlignment: MainAxisAlignment.center,
-          children: [
-            AnimatedContainer(
-              duration: Duration(seconds: 1),
-              curve: Curves.easeInOut,
-              child: Icon(
-                Icons.check_circle,
-                color: Color(0xFFFAAF84),
-                size: 100,
-              ),
-            ),
-            SizedBox(height: 20),
-            Text(
-              "Congratulations!",
-              style: TextStyle(
-                fontSize: 28,
-                fontWeight: FontWeight.bold,
-                color: Color(0xFFFAAF84),
-              ),
-            ),
-            SizedBox(height: 10),
-            Text(
-              "You have successfully signed up.",
-              style: TextStyle(fontSize: 16, color: Colors.grey[700]),
-            ),
-            SizedBox(height: 30),
-            ElevatedButton(
-              style: ElevatedButton.styleFrom(
-                backgroundColor: Color(0xFFFAAF84),
-                padding: EdgeInsets.symmetric(vertical: 15, horizontal: 30),
-                shape: RoundedRectangleBorder(
-                  borderRadius: BorderRadius.circular(15),
-                ),
-              ),
-              onPressed: () {
-                Navigator.pushAndRemoveUntil(
-                  context,
-                  MaterialPageRoute(
-                    builder: (_) => CounsellorSignInScreen(onSignOut: () async {
-                      await storage.deleteAll();
-                      Navigator.pushNamedAndRemoveUntil(context, '/', (route) => false);
-                      },),
-                  ),
-                  (route) => false,
-                );
-              },
-              child: Text(
-                "Sign In",
-                style: TextStyle(
-                  fontSize: 18,
-                  fontWeight: FontWeight.bold,
-                  color: Colors.white,
-                ),
-              ),
-            ),
-          ],
-        ),
-      ),
-    );
-  }
-}
