--- my_app/lib/screens/oldSignUpScreens/user_signup_data.dart
@@ -1,11 +1,0 @@
-class UserSignUpData {
-  String? username;
-  String? firstName;
-  String? lastName;
-  String? phoneNumber;
-  String? email;
-  String? password;
-  String? role = 'user';
-  List<String> userInterestedStates = [];
-  String? interestedCourse;
-}
