+++ my_app/lib/screens/signInScreens/get_user_details_step1.dart
@@ -1,31 +1,34 @@
 import 'package:flutter/material.dart';
-import 'package:my_app/screens/oldSignUpScreens/counselllor_signup_step3.dart';
-import 'package:my_app/screens/oldSignUpScreens/counsellor_signup_data.dart';
+import 'package:my_app/screens/signInScreens/get_user_details_step2.dart';
+import 'package:my_app/screens/signInScreens/user_details.dart';
 
-class CounsellorSignUpStep2 extends StatefulWidget {
-  final CounsellorSignUpData signUpData;
+class GetUserDetailsStep1 extends StatefulWidget {
+  final UserDetails userDetails;
+  final String userId;
+  final String jwtToken;
+  final String firebaseCustomToken;
 
-  CounsellorSignUpStep2({required this.signUpData});
+  final Future<void> Function() onSignOut;
+
+  GetUserDetailsStep1({
+    required this.userDetails,
+    required this.userId,
+    required this.jwtToken,
+    required this.firebaseCustomToken,
+    required this.onSignOut,
+  });
 
   @override
-  _CounsellorSignUpStep2State createState() => _CounsellorSignUpStep2State();
+  _GetUserDetailsStep1State createState() => _GetUserDetailsStep1State();
 }
 
-class _CounsellorSignUpStep2State extends State<CounsellorSignUpStep2> {
-  final List<String> _selectedExpertise = [];
+class _GetUserDetailsStep1State extends State<GetUserDetailsStep1> {
+  String? _selectedCourse;
 
   @override
   Widget build(BuildContext context) {
     return Scaffold(
-      backgroundColor: Colors.white,
-      appBar: AppBar(
-        backgroundColor: Color(0xFFFFE4B5), // Lighter shade of orange
-        title: Text(
-          "Sign Up",
-          style: TextStyle(color: Colors.white, fontWeight: FontWeight.bold),
-        ),
-        centerTitle: true,
-      ),
+      backgroundColor: Colors.white, // Set entire background to white
       body: SingleChildScrollView(
         child: Padding(
           padding: const EdgeInsets.all(16.0),
@@ -33,12 +36,14 @@ class _CounsellorSignUpStep2State extends State<CounsellorSignUpStep2> {
             crossAxisAlignment: CrossAxisAlignment.start,
             children: [
               Text(
-                "Choose Your Expertise",
+                "Choose Your Interested Course",
                 style: TextStyle(
-                  fontSize: 28,
+                  fontSize: 24,
                   fontWeight: FontWeight.bold,
-                  color: Color(0xFFFAAF84), // Slightly darker shade of orange
+                  color: Colors.black,
                 ),
+                maxLines: 1,
+                overflow: TextOverflow.ellipsis,
               ),
               SizedBox(height: 20),
               GridView.count(
@@ -55,27 +60,32 @@ class _CounsellorSignUpStep2State extends State<CounsellorSignUpStep2> {
                   _buildOption("OTHERS", Icons.more_horiz, "OTHERS"),
                 ],
               ),
-              SizedBox(height: 40),
+              SizedBox(height: 20),
               SizedBox(
                 width: double.infinity,
                 child: ElevatedButton(
                   style: ElevatedButton.styleFrom(
-                    backgroundColor: _selectedExpertise.isNotEmpty
-                        ? Color(0xFFFAAF84)
-                        : Colors.grey,
+                    backgroundColor: _selectedCourse != null
+                        ? Colors.green[300] // Light orange for enabled button
+                        : Colors.grey, // Grey for disabled button
                     padding: EdgeInsets.symmetric(vertical: 15),
                     shape: RoundedRectangleBorder(
                       borderRadius: BorderRadius.circular(15),
                     ),
                   ),
-                  onPressed: _selectedExpertise.isNotEmpty
+                  onPressed: _selectedCourse != null
                       ? () {
-                          widget.signUpData.expertise = _selectedExpertise;
+                          widget.userDetails.interestedCourse = _selectedCourse;
                           Navigator.push(
                             context,
                             MaterialPageRoute(
-                              builder: (context) => CounsellorSignUpStep3(
-                                  signUpData: widget.signUpData),
+                              builder: (context) => GetUserDetailsStep2(
+                                userDetails: widget.userDetails,
+                                userId: widget.userId,
+                                jwtToken: widget.jwtToken,
+                                firebaseCustomToken: widget.firebaseCustomToken,
+                                onSignOut: widget.onSignOut,
+                              ),
                             ),
                           );
                         }
@@ -98,41 +108,49 @@ class _CounsellorSignUpStep2State extends State<CounsellorSignUpStep2> {
   }
 
   Widget _buildOption(String label, IconData icon, String value) {
-    final isSelected = _selectedExpertise.contains(value);
+    final isSelected = _selectedCourse == value;
     return GestureDetector(
       onTap: () {
         setState(() {
-          if (isSelected) {
-            _selectedExpertise.remove(value);
-          } else {
-            _selectedExpertise.add(value);
-          }
+          _selectedCourse = value;
         });
       },
       child: Container(
+        width: MediaQuery.of(context).size.width *
+            0.1, // Smaller width (60% reduction)
+        height: MediaQuery.of(context).size.width *
+            0.1, // Smaller height (60% reduction)
         decoration: BoxDecoration(
-          color: isSelected ? Color(0xFFFAAF84).withOpacity(0.8) : Color(0xFFFFF8EE),
+          color: isSelected
+              ? Color(0xFFFFF3E0) // Light orange background for selected
+              : Colors.white, // White for unselected
           borderRadius: BorderRadius.circular(15),
-          border: Border.all(
-            color: isSelected ? Color(0xFFFAAF84) : Colors.grey.shade300,
-            width: 2,
-          ),
+          boxShadow: [
+            BoxShadow(
+              color: Colors.grey.withOpacity(0.3), // Subtle shadow
+              spreadRadius: 1,
+              blurRadius: 6,
+              offset: Offset(0, 2),
+            ),
+          ],
         ),
         child: Column(
           mainAxisAlignment: MainAxisAlignment.center,
           children: [
             Icon(
               icon,
-              size: 40,
-              color: isSelected ? Colors.white : Color(0xFFFAAF84),
+              size: 40, // Reduced size (50% of previous)
+              color: isSelected
+                  ? Color.fromARGB(255, 237, 110, 36)
+                  : Colors.black, // Orange for selected, black for unselected
             ),
             SizedBox(height: 10),
             Text(
               label,
               style: TextStyle(
-                fontSize: 16,
+                fontSize: 14, // Slightly smaller font
                 fontWeight: FontWeight.bold,
-                color: isSelected ? Colors.white : Color(0xFFFAAF84),
+                color: isSelected ? Color(0xFFFAAF84) : Colors.black,
               ),
             ),
           ],
