+++ my_app/lib/screens/signInScreens/privacy_page.dart
@@ -1,0 +1,15 @@
+import 'package:flutter/material.dart';
+
+class PrivacyPage extends StatelessWidget {
+  @override
+  Widget build(BuildContext context) {
+    return Scaffold(
+      appBar: AppBar(
+        title: Text('Privacy Policy'),
+        backgroundColor: Colors.white,
+      ),
+        backgroundColor: Colors.white,
+      body: Center(child: Text('This is the Privacy Policy page.')),
+    );
+  }
+}
